ShareekBackend/DepartmentApp/serializers.py
from rest_framework import serializers
from DepartmentApp.models import Departments

class DepartmentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Departments
        fields = ('DepartmentId',
                  'DepartmentName')
